Change type of all spatial values from int to float; Use round for float->int casting

+++ STRM/src/main/cpp/strm/include/strm/impl/models/MnnYoloV4Classifier.hpp
@@ -26,7 +26,7 @@ class MnnYoloV4Classifier : public Classifier {
 
   const float* getClassConfidences(const int i) const override;
 
-  Rect reconstructBox(float x, float y, float w, float h, int imageWidth, int imageHeight) override;
+  Rect reconstructBox(float x, float y, float w, float h, float imageWidth, float imageHeight) override;
 
   MNN::Interpreter* interpreter;
   MNN::Session* session;
